refactor: move pr manager into service layer (#68)
